refactor(types): extract schema types to lib/osm-tagging-schema
Prepare codebase for future external schema package (@gander-schemas/osm-tagging-schema):

- Create lib/osm-tagging-schema/ with unified schema data types
- Move GeometryType, FieldType, Field, Preset, PresetCategory, DeprecatedTag,
  SchemaData, TagIndex and related types to lib/osm-tagging-schema/types.ts
- Add PresetDefaults type for cleaner type definitions
- Update src/types/index.ts to re-export schema types from lib
- Keep MCP-specific types (SchemaLoaderConfig, TagValidationResult) in src/types
- Update tsconfig.json to include lib/ in compilation (rootDir: ".")
- Update package.json paths for dist/src/ output structure
- Fix path resolution in src/index.ts and src/version.ts for both
  tsx (source) and node (compiled) execution contexts

This prepares the codebase for:
1. Future extraction to @gander-schemas/osm-tagging-schema package
2. Multiple data sources that can be merged into unified SchemaData
3. Pre-processed and optimized schema data imports
